Changing the way the engine works
While loops are a nightmare to be paused and wait for user input. We
need a more static approach. Currently InputProvider has subscribers
(methods located within the class itself ) In these methods the new
"loop" is a method that updates the field. When an event is fired from
the UI the subscribers are notifyed and call the engine.UpdateField
method is called. Not very good I know but its the best working solution
so fat.